requirements.txt
Django==5.2.7
asgiref==3.10.0
kafka-python==2.2.15
psycopg2-binary==2.9.11
sqlparse==0.5.3
pillow~=12.0.0